requirements.txt
asgiref==3.3.4
certifi==2021.10.8
charset-normalizer==2.0.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.3
django-cors-headers==3.7.0
django-import-export==2.6.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
et-xmlfile==1.1.0
idna==3.3
import-export==0.2.67.dev6
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
PyJWT==1.7.1
pytz==2021.1
PyYAML==6.0
requests==2.26.0
sqlparse==0.4.1
tablib==3.1.0
urllib3==1.26.7
xlrd==2.0.1
xlwt==1.3.0
